requirements.txt
fpdf2
qrcode